Extracted Telemetry out of Globals
