+++ src/main/java/org/jabref/gui/Globals.java
@@ -1,10 +1,5 @@
 package org.jabref.gui;
 
-import java.util.Optional;
-import java.util.UUID;
-
-import javafx.stage.Screen;
-
 import org.jabref.architecture.AllowedToUseAwt;
 import org.jabref.gui.keyboard.KeyBindingRepository;
 import org.jabref.gui.remote.CLIMessageHandler;
@@ -19,14 +14,9 @@ import org.jabref.logic.remote.RemotePreferences;
 import org.jabref.logic.remote.server.RemoteListenerServerManager;
 import org.jabref.logic.util.BuildInfo;
 import org.jabref.model.entry.BibEntryTypesManager;
-import org.jabref.model.strings.StringUtil;
 import org.jabref.model.util.FileUpdateMonitor;
 import org.jabref.preferences.JabRefPreferences;
 
-import com.google.common.base.StandardSystemProperty;
-import com.microsoft.applicationinsights.TelemetryClient;
-import com.microsoft.applicationinsights.TelemetryConfiguration;
-import com.microsoft.applicationinsights.telemetry.SessionState;
 import kong.unirest.Unirest;
 
 /**
@@ -77,7 +67,6 @@ public class Globals {
     private static ThemeManager themeManager;
 
     private static DefaultFileUpdateMonitor fileUpdateMonitor;
-    private static TelemetryClient telemetryClient;
 
     private Globals() {
     }
@@ -119,37 +108,12 @@ public class Globals {
     public static void startBackgroundTasks() {
         // TODO Currently deactivated due to incompatibilities in XML
       /*  if (Globals.prefs.getTelemetryPreferences().shouldCollectTelemetry() && !GraphicsEnvironment.isHeadless()) {
-            startTelemetryClient();
+            Telemetry.start(prefs.getTelemetryPreferences());
         } */
         RemotePreferences remotePreferences = prefs.getRemotePreferences();
         if (remotePreferences.useRemoteServer()) {
             Globals.REMOTE_LISTENER.openAndStart(new CLIMessageHandler(prefs, fileUpdateMonitor, entryTypesManager), remotePreferences.getPort());
         }
-    }
-
-    private static void stopTelemetryClient() {
-        getTelemetryClient().ifPresent(client -> {
-            client.trackSessionState(SessionState.End);
-            client.flush();
-        });
-    }
-
-    private static void startTelemetryClient() {
-        TelemetryConfiguration telemetryConfiguration = TelemetryConfiguration.getActive();
-        if (!StringUtil.isNullOrEmpty(Globals.BUILD_INFO.azureInstrumentationKey)) {
-            telemetryConfiguration.setInstrumentationKey(Globals.BUILD_INFO.azureInstrumentationKey);
-        }
-        telemetryConfiguration.setTrackingIsDisabled(!Globals.prefs.getTelemetryPreferences().shouldCollectTelemetry());
-        telemetryClient = new TelemetryClient(telemetryConfiguration);
-        telemetryClient.getContext().getProperties().put("JabRef version", Globals.BUILD_INFO.version.toString());
-        telemetryClient.getContext().getProperties().put("Java version", StandardSystemProperty.JAVA_VERSION.value());
-        telemetryClient.getContext().getUser().setId(Globals.prefs.getTelemetryPreferences().getUserId());
-        telemetryClient.getContext().getSession().setId(UUID.randomUUID().toString());
-        telemetryClient.getContext().getDevice().setOperatingSystem(StandardSystemProperty.OS_NAME.value());
-        telemetryClient.getContext().getDevice().setOperatingSystemVersion(StandardSystemProperty.OS_VERSION.value());
-        telemetryClient.getContext().getDevice().setScreenResolution(Screen.getPrimary().getVisualBounds().toString());
-
-        telemetryClient.trackSessionState(SessionState.Start);
     }
 
     public static void shutdownThreadPools() {
@@ -161,11 +125,7 @@ public class Globals {
     }
 
     public static void stopBackgroundTasks() {
-        stopTelemetryClient();
+        Telemetry.shutdown();
         Unirest.shutDown();
-    }
-
-    public static Optional<TelemetryClient> getTelemetryClient() {
-        return Optional.ofNullable(telemetryClient);
     }
 }
