+++ src/main/java/org/jabref/gui/importer/actions/OpenDatabaseAction.java
@@ -14,10 +14,10 @@ import java.util.Optional;
 import javax.swing.undo.UndoManager;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.Globals;
 import org.jabref.gui.JabRefFrame;
 import org.jabref.gui.LibraryTab;
 import org.jabref.gui.StateManager;
+import org.jabref.gui.Telemetry;
 import org.jabref.gui.actions.SimpleCommand;
 import org.jabref.gui.autosaveandbackup.BackupManager;
 import org.jabref.gui.dialogs.BackupUIManager;
@@ -256,7 +256,7 @@ public class OpenDatabaseAction extends SimpleCommand {
     }
 
     private void trackOpenNewDatabase(LibraryTab libraryTab) {
-        Globals.getTelemetryClient().ifPresent(client -> client.trackEvent(
+        Telemetry.getTelemetryClient().ifPresent(client -> client.trackEvent(
                 "OpenNewDatabase",
                 Map.of(),
                 Map.of("NumberOfEntries", (double) libraryTab.getBibDatabaseContext().getDatabase().getEntryCount())));
